Move graph_resource fixture to ndbfixtures
