+++ nucliadb/tests/nucliadb/integration/test_ask.py
@@ -48,10 +48,7 @@ from nucliadb_models.search import (
     RagStrategies,
     SyncAskResponse,
 )
-from nucliadb_protos.utils_pb2 import Relation, RelationMetadata, RelationNode
-from nucliadb_protos.writer_pb2 import BrokerMessage
-from tests.utils import inject_message
-from tests.utils.dirty_index import wait_for_sync
+from nucliadb_protos.utils_pb2 import RelationNode
 
 
 @pytest.fixture(scope="function", autouse=True)
@@ -115,92 +112,6 @@ async def resource(nucliadb_writer: AsyncClient, standalone_knowledgebox: str):
     rid = resp.json()["uuid"]
 
     yield rid
-
-
-@pytest.fixture
-async def graph_resource(nucliadb_writer: AsyncClient, nucliadb_ingest_grpc, standalone_knowledgebox):
-    resp = await nucliadb_writer.post(
-        f"/kb/{standalone_knowledgebox}/resources",
-        json={
-            "title": "Knowledge graph",
-            "slug": "knowledgegraph",
-            "summary": "Test knowledge graph",
-            "texts": {
-                "inception1": {"body": "Christopher Nolan directed Inception. Very interesting movie."},
-                "inception2": {"body": "Leonardo DiCaprio starred in Inception."},
-                "inception3": {"body": "Joseph Gordon-Levitt starred in Inception."},
-                "leo": {"body": "Leonardo DiCaprio is a great actor. DiCaprio started acting in 1989."},
-            },
-        },
-    )
-    assert resp.status_code == 201
-    rid = resp.json()["uuid"]
-
-    nodes = {
-        "nolan": RelationNode(
-            value="Christopher Nolan", ntype=RelationNode.NodeType.ENTITY, subtype="DIRECTOR"
-        ),
-        "inception": RelationNode(
-            value="Inception", ntype=RelationNode.NodeType.ENTITY, subtype="MOVIE"
-        ),
-        "leo": RelationNode(
-            value="Leonardo DiCaprio", ntype=RelationNode.NodeType.ENTITY, subtype="ACTOR"
-        ),
-        "dicaprio": RelationNode(value="DiCaprio", ntype=RelationNode.NodeType.ENTITY, subtype="ACTOR"),
-        "levitt": RelationNode(
-            value="Joseph Gordon-Levitt", ntype=RelationNode.NodeType.ENTITY, subtype="ACTOR"
-        ),
-    }
-    edges = [
-        Relation(
-            relation=Relation.RelationType.ENTITY,
-            source=nodes["nolan"],
-            to=nodes["inception"],
-            relation_label="directed",
-            metadata=RelationMetadata(
-                # Set this field id as int enum value since this is how legacy relations reported paragraph_id
-                paragraph_id=rid + "/4/inception1/0-37",
-                data_augmentation_task_id="my_graph_task_id",
-            ),
-        ),
-        Relation(
-            relation=Relation.RelationType.ENTITY,
-            source=nodes["leo"],
-            to=nodes["inception"],
-            relation_label="starred",
-            metadata=RelationMetadata(
-                paragraph_id=rid + "/t/inception2/0-39",
-                data_augmentation_task_id="my_graph_task_id",
-            ),
-        ),
-        Relation(
-            relation=Relation.RelationType.ENTITY,
-            source=nodes["levitt"],
-            to=nodes["inception"],
-            relation_label="starred",
-            metadata=RelationMetadata(
-                paragraph_id=rid + "/t/inception3/0-42",
-                data_augmentation_task_id="",
-            ),
-        ),
-        Relation(
-            relation=Relation.RelationType.ENTITY,
-            source=nodes["leo"],
-            to=nodes["dicaprio"],
-            relation_label="analogy",
-            metadata=RelationMetadata(
-                paragraph_id=rid + "/t/leo/0-70",
-                data_augmentation_task_id="my_graph_task_id",
-            ),
-        ),
-    ]
-    bm = BrokerMessage()
-    bm.uuid = rid
-    bm.kbid = standalone_knowledgebox
-    bm.user_relations.relations.extend(edges)
-    await inject_message(nucliadb_ingest_grpc, bm)
-    await wait_for_sync()
-    return rid
 
 
 @pytest.mark.deploy_modes("standalone")
